+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+# 0.23.0 - 2024-04-18
+
 - Update winston logger to push json format instead of raw text and enable colors
 - Update typing for `/v1/troubleshoot`, now it will behaved like an action route